Add file test for IGNORE_GLOBAL_HOOKS
This change will add a file check test for the IGNORE_GLOBAL_HOOKS
variable.  This will prevent an error from the subsequent `cat` command
if the file does not exist.  The error from `cat` is innocuous, but
suppressing it will reduce any confusion for users who see an error in
the commit messages should the file IGNORE_GLOBAL_HOOKS not exist.